v2/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.1.4
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-filter==22.1
django-tables2==2.4.1
djangorestframework==3.14.0
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
